Export auth interface to allow external implementations
The original driver auth interface and implementations have been moved into their own packages and exported. There are two implementations within the driver, NTML and Windows kerberos.
As before the default on windows is the winsspi package and on, on Linux, NTLM.

The consuming application can now override this at runtime by calling mssql.SetAuthProvider(authProvider)

This allows the application to provide a custom implementation of the auth.Provider interface.
auth.Provider is in turn responsible for creating an instance of auth.Auth for the required authentication mechanism. windows, NTML etc.