Fixed #93: servername was not supplied to TLS cfg
A change in fd44003 moved the code for setting p.Host down in the parse
function, resulting in an empty p.Host value at the time of the creation
of p.TLSConfig. I extracted the TLS parsing into a separate function and
moved the TLSConfig creation to the end of the parsing function.