Fix: ExampleLastInsertId() -> ExampleResult_LastInsertId()
The new tests analyzer reports common mistakes in declarations of
tests, fuzzers, benchmarks, and examples in test packages, such as
malformed names, incorrect signatures, or examples that document
non-existent identifiers. Some of these mistakes may cause tests
not to run. This analyzer is among the subset of analyzers that
are run by go test.

See:
https://go.dev/doc/go1.24#vet
https://go.dev/blog/examples#example-function-names